Fixed all?(most) errors if user enters edit or delete urls without existing id's and design changes/cleaning up

+++ main.py
@@ -1,11 +1,13 @@
 from wtforms import StringField, EmailField, PasswordField, SubmitField
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
-from flask import Flask, render_template, url_for, request, redirect
-from wtforms.validators import DataRequired
+from flask import Flask, render_template, url_for, request, redirect, flash
+from wtforms.validators import DataRequired, Length
 from flask_sqlalchemy import SQLAlchemy
 from flask_bootstrap import Bootstrap5
 from flask_wtf import FlaskForm
 # from sqlalchemy import String, Integer
+
+#  dont know if bootstrap and css fix is possible..
 # TODO: Sort the passwords on the user page alphabetically
 class Base(DeclarativeBase):
     pass
@@ -30,10 +32,11 @@ class Passwords(db.Model):
 
 # TODO: Implement generative passwords
 # The form to add a website, email and password
+# TODO: Handled error of strings entered being too big to display for the css with code for now
 class EntryForm(FlaskForm):
-    website = StringField(name="website", validators=[DataRequired()])
-    email = EmailField(name="email", validators=[DataRequired()])
-    password = PasswordField(name="password", validators=[DataRequired()])
+    website = StringField(name="website", validators=[DataRequired(), Length(max=40)])
+    email = EmailField(name="email", validators=[DataRequired(), Length(max=40)])
+    password = PasswordField(name="password", validators=[DataRequired(), Length(max=40)])
     submit = SubmitField(name="submit", validators=[DataRequired()])
 
 
@@ -43,14 +46,27 @@ with app.app_context():
 
 @app.route("/")
 def home():
-
     return render_template("index.html", )
+
+@app.route("/manage_passwords")
+def manage_passwords():
+    passwords = db.session.scalars(db.select(Passwords)).all()
+    return render_template("user.html", passwords=passwords)
+
 
 # TODO: ADD a check to make sure the user isnt adding the same website twice, is unqiue will catch and crash
 @app.route('/add', methods=["GET", "POST"])
 def add():
     form = EntryForm()
+    # TODO: stop someone from entering the same website
     if form.validate_on_submit():
+
+        # TODO: Replace with verifying function, same for edit and delete
+        check_website = [item.website for item in db.session.scalars((db.select(Passwords))).all()]
+        if request.form.get('website') in check_website:
+            flash("Password for website already exists")
+            return render_template("add.html", form=form)
+
         new_entry = Passwords(website= request.form.get('website'),
                               email= request.form.get('email'),
                               password = request.form.get('password'),
@@ -61,16 +77,12 @@ def add():
     return render_template("add.html", form=form)
 
 
-@app.route("/user_profile")
-def user_profile():
-    passwords = db.session.scalars(db.select(Passwords)).all()
-    return render_template("user.html", passwords=passwords)
-
 # TODO: This can collapse into the "add" route instead by passing the id and having and "if id:" statement in add.
 @app.route("/edit", methods=["GET", "POST"])
 def edit():
-    # to protect against entering the edit url without an id
-    if request.args.get('id'):
+    # Checks ids in database to see if requested password to edit exists, to protect against manually enter id in url
+    check_id = [str(item.id) for item in db.session.scalars((db.select(Passwords))).all()]
+    if request.args.get('id') in check_id:
         entry = db.session.scalar((db.select(Passwords).where(Passwords.id == request.args.get('id'))))
         form = EntryForm(
             website=entry.website,
@@ -82,16 +94,19 @@ def edit():
             entry.email = form.email.data
             entry.password = form.password.data
             db.session.commit()
-            return redirect(url_for('user_profile'))
-        return render_template("edit.html", form=form)
+            return redirect(url_for('manage_passwords'))
+        return render_template("add.html", form=form)
     return redirect(url_for("add"))
 
-
+# TODO: For editing too, make a function that checks a result exists, so we can centrally deal with failed requests
 @app.route("/delete")
 def delete():
-    db.session.delete(db.session.scalar((db.select(Passwords).where(Passwords.id == request.args.get('id')))))
-    db.session.commit()
-    return redirect(url_for("user_profile"))
+    check_id = [str(item.id) for item in db.session.scalars((db.select(Passwords))).all()]
+    if request.args.get('id') in check_id:
+        db.session.delete(db.session.scalar((db.select(Passwords).where(Passwords.id == request.args.get('id')))))
+        db.session.commit()
+        return redirect(url_for("manage_passwords"))
+    return redirect(url_for("manage_passwords"))
 
 if __name__ == "__main__":
     app.run(debug=True)
